+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+## 0.0.3
+- basic ncurses mode
+
 ## 0.0.2
-- Ready in basic report mode
+- basic report mode
 
 ## 0.0.1
 
-- Initial version
+- initial version